requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6
starlette==0.27.0
itsdangerous==2.1.2

# Google OAuth and authentication
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
authlib==1.3.0

# JWT handling
PyJWT==2.8.0
cryptography==41.0.7

# Google Cloud services
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.17.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# CORS support
python-jose[cryptography]==3.3.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0